+++ cli/Cargo.toml
@@ -7,5 +7,6 @@ edition = "2018"
 [dependencies]
 radicle_registry_client = { path = "../client" }
 
+futures01 = { package = "futures", version = "0.1" }
 pretty_env_logger = "0.3.1"
 structopt = "0.3"
